requirements.txt
opencv-python
face-recognition
numpy
pandas
customtkinter
pillow
cloudinary
